==> experiments/calibrate/04-supp-notes.ipynb <==
%load_ext lab_black
import pandas as pd
import numpy as np
import matplotlib
import matplotlib.pyplot as plt
from scipy import stats
import statsmodels.api as sm
from scipy.stats import pearsonr

matplotlib.rcParams["font.family"] = ["Liberation Sans"]

# Question we seek to answer
1. For LDL in all populations, $R^2$ increase with PC1, but coverage is ok.
2. For LDL in WB populations, $R^2$ increase with sex, but coverage is ok.
3. For BMI in WB populations, $R^2$ increase with sex, but coverage is ok.

the observation in LDL is due to the var(x). because R2 is a function of var(x) even if var(e) is the same across all covariates. write this into supp. notes / methods. posterior check should also be written into supp. notes. question: does this still happen if i do not include PGS x PC1 into the model? (compare R2 between w/o PGSxPC1 and w. PGSxPC1, or directly compare the figure) we should do the comparison just by varying this. This can also be added to the supp. notes.

For LDL, the ground truth is that the residual noise are similar across PC1 groups. But the slope is different across groups.

If PGS (PC1 + 1) is modeled, then it will capture that.

Otherwise if variable slope, the residual noise are still similar (maybe a bit larger than the scenraio where variable slope is not modeled)

more

- $R^2$ between polygenic score and trait is $R^2(y, \widehat{y})$. 
- Under a model of $y = \widehat{y} + e$. Interpretation being that the true phenotype floats around the point predictions. Or we understand this as: for all individuals with the same point predictions, the phenotypes follow a normal distribution around point predictions.
- $R^2$ is $R^2(\widehat{y} + e, \widehat{y}) = \frac{\text{Var}[\widehat{y}]}{\text{Var}[\widehat{y}] + \text{Var}[e]}$. Therefore, under this model, even when prediction interval length is held constant, i.e. $\text{Var}[e]$ is constant, variable $R^2$ will derive from variable $\text{Var}[\widehat{y}]$ across groups.
- Such variable $\text{Var}[\widehat{y}]$ can come from multiple factors. One of important sources is because has variable slope.

We verify this in several examples. And we additionally provide simulations showcasing predictive checks that this can happen even when  

# setup (so this notebook can be run for trait-group repeatedly)
trait = "LDL"
group = "all"

# This hypothesis can be validated using predictive checks

def predictive_check(trait, group, covar):
    dict_r2 = dict()
    for method in ["mean", "mean+var"]:
        r2_list = []
        for seed in range(1, 31):
            df = pd.read_csv(
                f"out/model/{trait}-{group}/{seed}-{method}.test_info.tsv",
                sep="\t",
                index_col=0,
            )
            df_params = pd.read_csv(
                f"out/model/{trait}-{group}/{seed}-{method}.param.tsv",
                sep="\t",
                index_col=0,
            )
            if len(df[covar].unique()) > 5:
                df["q"] = pd.qcut(df[covar], q=5).cat.codes
            else:
                df["q"] = df[covar].values

            df["pheno_sim"] = np.random.normal(
                loc=df["cal_pred"], scale=df["cal_predstd"]
            )
            r2 = df.groupby("q").apply(
                lambda dfq: stats.pearsonr(dfq["cal_pred"], dfq["pheno_sim"])[0] ** 2
            )
            r2_list.append(r2)
        r2_list = pd.concat(r2_list, axis=1).T
        dict_r2[method] = r2_list

    fig, axes = plt.subplots(figsize=(6, 3), dpi=150, ncols=2, sharey=True)
    for i, method in enumerate(["mean", "mean+var"]):
        ax = axes[i]
        ax.boxplot(dict_r2[method], notch=True)
        ax.set_title(method, fontsize=11)
        ax.set_xlabel(covar)
    fig.suptitle(f"{trait}, {group}", y=0.9)
    fig.tight_layout()
    fig.show()

for covar in ["AGE", "SEX", "PC1"]:
    predictive_check(trait="LDL", group="all", covar=covar)

# This can also be validated using numerical evidence

- Stratify regression analysis between $y$ and $\widehat{y}$ across age/sex/PC1
- Stratify regression analysis between $y$ and individual covariates age/sex/PC1 (print top covariates)

def regression(trait, group, covar):
    method = "mean+var"
    df_stats_list = []
    for seed in range(1, 31):
        df = pd.read_csv(
            f"out/model/{trait}-{group}/{seed}-{method}.test_info.tsv",
            sep="\t",
            index_col=0,
        )
        if len(df[covar].unique()) > 5:
            df["q"] = pd.qcut(df[covar], q=5).cat.codes
        else:
            df["q"] = df[covar].values

        df_stats = {
            "r2": [],
            "std(resid)": [],
            "std(fitted_resid)": [],
            "std(pred)": [],
            "std(y)": [],
        }

        for i, (q, dfq) in enumerate(df.groupby("q")):
            x, y = dfq["cal_pred"], dfq["pheno"]
            model = sm.OLS(endog=y, exog=sm.add_constant(x)).fit()
            df_stats["r2"].append(model.rsquared)
            df_stats["std(resid)"].append(np.std(model.resid))
            df_stats["std(fitted_resid)"].append(np.std(y - x))
            df_stats["std(pred)"].append(np.std(x))
            df_stats["std(y)"].append(np.std(y))
        df_stats = pd.DataFrame(df_stats)
        df_stats_list.append(df_stats)

    df_stats = pd.DataFrame(
        np.mean(df_stats_list, axis=0),
        index=df_stats_list[0].index,
        columns=df_stats_list[0].columns,
    )
    return df_stats

trait, group = "LDL", "all"
seed = 1
method = "mean+var"

df_params = pd.read_csv(
    f"out/model/{trait}-{group}/{seed}-{method}.param.tsv",
    sep="\t",
    index_col=0,
)

df_params

for covar in ["PC1", "AGE", "SEX"]:
    print(f"# {covar}")
    display(regression("LDL", "all", covar))

for covar in ["PC1", "AGE", "SEX"]:
    print(f"# {covar}")
    display(regression("height", "all", covar))
